Split up subscriber
