+++ src/mobot_client/drop_session.py
@@ -104,14 +104,14 @@ class BaseDropSession:
             ChatStrings.HELP
         )
 
-    def handle_cancel(self, message, drop_session: DropSession):
+    def handle_cancel(self, drop_session: DropSession):
         drop_session.state = SessionState.CANCELLED
         drop_session.save()
         self.messenger.log_and_send_message(
             ChatStrings.SESSION_CANCELLED
         )
 
-    def handle_privacy_policy_request(self, message, drop_session: DropSession):
+    def handle_privacy_policy_request(self, drop_session: DropSession):
         privacy_policy_url = drop_session.drop.store.privacy_policy_url
         self.messenger.log_and_send_message(
             ChatStrings.PRIVACY_POLICY.format(url=privacy_policy_url),
