+++ src/mobot_client/item_drop_session.py
@@ -1,24 +1,22 @@
 # Copyright (c) 2021 MobileCoin. All rights reserved.
 
-import os
 from decimal import Decimal
 
-import mobilecoin as mc
 import googlemaps
 import pytz
 
 from django.conf import settings
 
+from mobot_client.core.context import ChatContext
 from mobot_client.drop_session import BaseDropSession
 from mobot_client.models import (
     DropSession,
-    Customer,
     Drop,
     CustomerStorePreferences,
     Order,
     Sku,
     CustomerDropRefunds, SessionState, OrderStatus,
-    Item, OutOfStockException,
+    OutOfStockException,
 )
 from mobot_client.chat_strings import ChatStrings
 from mobot_client.models.messages import Message
@@ -42,13 +40,13 @@ class ItemDropSession(BaseDropSession):
         elif drop_session.state == SessionState.SHIPPING_INFO_CONFIRMATION:
             self.handle_item_drop_session_shipping_confirmation(message, drop_session)
         elif drop_session.state == SessionState.ALLOW_CONTACT_REQUESTED:
-            self.handle_item_drop_session_allow_contact_requested(message, drop_session)
+            self.handle_item_drop_session_allow_contact_requested(drop_session)
         elif drop_session.state == SessionState.READY:
-            self.show_catalog_and_display_payment_instructions(message, drop_session)
+            self.show_catalog_and_display_payment_instructions(drop_session)
 
     def handle_item_payment(self, amount_paid_mob: Decimal, drop_session: DropSession):
         customer = drop_session.customer
-        payment_succeeded = self.payments.handle_item_payment(customer.phone_number.as_e164, customer, amount_paid_mob, drop_session)
+        payment_succeeded = self.payments.handle_item_payment(amount_paid_mob, drop_session)
         if payment_succeeded:
             message_to_send = (
                     ChatStrings.WAITING_FOR_SIZE_PREFIX + ChatStrings.get_options(list(drop_session.drop.item.skus.all()),
@@ -67,7 +65,7 @@ class ItemDropSession(BaseDropSession):
                 ChatStrings.ITEM_OPTION_HELP + "\n\n" + ChatStrings.get_options(drop_session.drop.item.skus, capitalize=True)
             )
         elif message.text.lower() == "privacy":
-            self.handle_privacy_policy_request(message, drop_session)
+            self.handle_privacy_policy_request(drop_session)
             self.show_catalog(message, drop_session)
         elif message.text.lower() == "chart" or message.text.lower() == 'info':
             drop_item = drop_session.drop.item
@@ -95,14 +93,14 @@ class ItemDropSession(BaseDropSession):
             sku: Sku = drop_session.drop.item.skus.filter(identifier__iexact=message.text).first()
             if not sku:
                 message_to_send = f"{message.text} is not an available size"
-                self.show_catalog(message, drop_session, message_to_send)
+                self.show_catalog(drop_session, message_to_send)
             else:
                 try:
                     # Atomically order the item
                     order = sku.order(drop_session)
                     self.logger.info(f"Successfully created order: {order}")
                 except OutOfStockException:
-                    self.show_catalog(message, drop_session, ChatStrings.ITEM_SOLD_OUT)
+                    self.show_catalog(drop_session, ChatStrings.ITEM_SOLD_OUT)
                 else:
                     drop_session.state = SessionState.WAITING_FOR_NAME
                     drop_session.save()
@@ -117,13 +115,13 @@ class ItemDropSession(BaseDropSession):
                 ChatStrings.ITEM_HELP
             )
         elif message.text.lower() == "cancel":
-            self.handle_cancel()
+            self.handle_cancel(drop_session)
         elif message.text.lower() == "pay":
             self.messenger.log_and_send_message(
                 ChatStrings.PAY.format(amount=drop_session.drop.item.price_in_mob.normalize()),
             )
         elif message.text.lower() == "privacy":
-            self.handle_privacy_policy_request(message, drop_session)
+            self.handle_privacy_policy_request(drop_session)
         elif message.text.lower() == "info":
             drop_item = drop_session.drop.item
 
@@ -147,7 +145,7 @@ class ItemDropSession(BaseDropSession):
                 ChatStrings.ITEM_HELP_SHORT
             )
         # Re-display available sizes and request payment
-        self.show_catalog_and_display_payment_instructions(message, drop_session)
+        self.show_catalog_and_display_payment_instructions(drop_session)
         drop_session.save()
 
     def handle_cancel_and_refund(self, message, drop_session, order):
@@ -177,7 +175,9 @@ class ItemDropSession(BaseDropSession):
                 if TYPE == "country" and component["short_name"] != drop_session.drop.country_code_restriction:
                     return False
 
-    def validated_address(self, message, drop_session: DropSession):
+    def validated_address(self, drop_session: DropSession):
+        ctx = ChatContext.get_current_context()
+        message = ctx.message
         address = self.gmaps.geocode(message.text, region=drop_session.drop.country_code_restriction)
         order = drop_session.order
         if len(address) == 0:
@@ -199,7 +199,7 @@ class ItemDropSession(BaseDropSession):
             self.messenger.log_and_send_message(
                 ChatStrings.MISSING_ORDER
             )
-        elif address := self.validated_address(message, drop_session):
+        elif address := self.validated_address(drop_session):
             order.shipping_address = address[0]["formatted_address"]
             order.save()
 
@@ -238,7 +238,7 @@ class ItemDropSession(BaseDropSession):
                 ChatStrings.ADDRESS_REQUEST
             )
 
-    def confirm_order(self, message, drop_session: DropSession):
+    def confirm_order(self, drop_session: DropSession):
         order = drop_session.order
         order.status = OrderStatus.CONFIRMED
         order.save()
@@ -263,7 +263,7 @@ class ItemDropSession(BaseDropSession):
             ),
         )
 
-    def handle_item_drop_session_shipping_confirmation(self, message, drop_session: DropSession):
+    def handle_item_drop_session_shipping_confirmation(self, message: Message, drop_session: DropSession):
         if not drop_session.order:
             self.messenger.log_and_send_message(
                 ChatStrings.MISSING_ORDER
@@ -275,7 +275,7 @@ class ItemDropSession(BaseDropSession):
                 ChatStrings.NAME_REQUEST
             )
         elif message.text.lower() == "privacy":
-            self.handle_privacy_policy_request(message, drop_session)
+            self.handle_privacy_policy_request(drop_session)
         elif message.text.lower() == "c" or message.text.lower() == "cancel" or message.text.lower() == "refund":
             self.handle_cancel_and_refund(message, drop_session, drop_session.order)
         elif message.text.lower() != "yes" and message.text.lower() != "y":
@@ -285,11 +285,12 @@ class ItemDropSession(BaseDropSession):
                 )
             )
         else:
-            self.confirm_order(message, drop_session)
+            self.confirm_order(drop_session)
         # Finish off by offering further contact
-        self.handle_item_drop_session_allow_contact_requested(message, drop_session)
+        self.handle_item_drop_session_allow_contact_requested(drop_session)
 
-    def handle_item_drop_session_allow_contact_requested(self, message, drop_session):
+    def handle_item_drop_session_allow_contact_requested(self, drop_session: DropSession):
+        message = ChatContext.get_current_context().message
         if message.text.lower() == "n" or message.text.lower() == "no":
             customer_store_prefs = CustomerStorePreferences(
                 customer=drop_session.customer, store=self.store, allows_contact=False
@@ -309,7 +310,7 @@ class ItemDropSession(BaseDropSession):
             )
             drop_session.state = SessionState.COMPLETED
         elif message.text.lower() == 'p' or message.text.lower() == "privacy" or message.text.lower() == "privacy policy":
-            self.handle_privacy_policy_request(message, drop_session)
+            self.handle_privacy_policy_request(drop_session)
         else:
             self.messenger.log_and_send_message(
                 ChatStrings.HELP
@@ -322,9 +323,8 @@ class ItemDropSession(BaseDropSession):
         )
         drop_session.state = SessionState.WAITING_FOR_PAYMENT
 
-    def show_catalog(self, message, drop_session: DropSession, message_to_send: str = ""):
+    def show_catalog(self, drop_session: DropSession, message_to_send: str = ""):
         drop = drop_session.drop
-        customer = drop_session.customer
         available_options = drop.item.skus.all()
         if not available_options.exists():
             self.messenger.log_and_send_message(
@@ -340,10 +340,10 @@ class ItemDropSession(BaseDropSession):
             )
             self.messenger.log_and_send_message(message_to_send)
 
-    def show_catalog_and_display_payment_instructions(self, message: Message, drop_session: DropSession):
-        customer = drop_session.customer
+    def show_catalog_and_display_payment_instructions(self, drop_session: DropSession):
+        ctx = ChatContext.get_current_context()
         drop = drop_session.drop
-        customer_payments_address = self.payments.get_payments_address(message.customer.phone_number.as_e164)
+        customer_payments_address = self.payments.get_payments_address(ctx.customer.phone_number.as_e164)
         if customer_payments_address is None:
             self.messenger.log_and_send_message(
                 ChatStrings.PAYMENTS_ENABLED_HELP.format(
@@ -357,11 +357,13 @@ class ItemDropSession(BaseDropSession):
                 store_description=drop.store.description,
                 item_description=drop.item.short_description
             )
-            self.show_catalog(message, drop_session, greeting_text)
+            self.show_catalog(drop_session, greeting_text)
             self.request_payment_from_customer(drop_session, drop_session.drop.item.price_in_mob)
             drop_session.save()
 
-    def handle_no_active_item_drop_session(self, customer, message, drop):
+    def handle_no_active_item_drop_session(self, drop: Drop):
+        ctx = ChatContext.get_current_context()
+        customer = ctx.customer
         new_drop_session, _ = DropSession.objects.get_or_create(
             customer=customer,
             drop=drop,
